backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
scikit-learn==1.3.2
scipy==1.11.4
numpy==1.25.2
pandas==2.1.3
requests==2.31.0
aiofiles==23.2.1
apscheduler==3.10.4
pyttsx3==2.90
pyclamd==0.4.0
python-magic-bin==0.4.14
psutil==5.9.6
pystray==0.19.5
Pillow==10.1.0
watchdog==3.0.0